+++ scripts/discrete_representation/octree/inner_growing_showcase.py
@@ -4,7 +4,7 @@ Showcase of inner growing voxelization.
 import random
 
 import volmdlr
-from volmdlr.core import VolumeModel
+from volmdlr.model import VolumeModel
 from volmdlr.discrete_representation import MatrixBasedVoxelization
 from volmdlr.primitives3d import Block