+++ scripts/showcases/casing.py
@@ -8,7 +8,7 @@ import math
 
 import plot_data
 
-from volmdlr.models import casing
+from volmdlr.models.casing import casing
 from volmdlr.primitives3d import Block
 
 casing._check_platform()